# backend/src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Siinader API",
        "description": "Documentación de la API usada para el sistema Siinader",
        "servers": [
            "http://localhost:3000"
        ]
    },
    "host": "localhost:3000",
    "basePath": "",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/api/estudiantes": {
            "get": {
                "tags": [
                    {
                        "name": "Estudiante",
                        "description": " API para los estudiantes"
                    }
                ],
                "summary": "Obtiene todos los estudiantes",
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa",
                        "schema": {
                            "$ref": "#/definitions/Estudiante"
                        }
                    },
                    "400": {
                        "description": "Datos enviados incorrectos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "post": {
                "tags": [
                    {
                        "name": "Estudiante",
                        "description": " API para los estudiantes"
                    }
                ],
                "summary": "Crea un estudiante",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "El estudiantes que se creará, no debe incluir id",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Estudiante"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa",
                        "schema": {
                            "$ref": "#/definitions/Estudiante"
                        }
                    },
                    "400": {
                        "description": "Datos enviados incorrectos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
        "/api/estudiantes/:id": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID del estudiante",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    {
                        "name": "Estudiante",
                        "description": " API para los estudiantes"
                    }
                ],
                "summary": "Obtiene el estudiante con el ID dado",
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa",
                        "schema": {
                            "$ref": "#/definitions/Estudiante"
                        }
                    },
                    "400": {
                        "description": "Datos enviados incorrectos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    {
                        "name": "Estudiante",
                        "description": " API para los estudiantes"
                    }
                ],
                "summary": "Actualiza al estudiante con el ID dado",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "description": "El estudiante se actulizará, no debe incluir CI",
                        "schema": {
                            "$ref": "#/definitions/Estudiante"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa",
                        "schema": {
                            "$ref": "#/definitions/Estudiante"
                        }
                    },
                    "400": {
                        "description": "Datos enviados incorrectos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    {
                        "name": "Estudiante",
                        "description": " API para los estudiantes"
                    }
                ],
                "summary": "Elimina al estudiante con el ID dado",
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa",
                        "schema": {
                            "$ref": "#/definitions/Estudiante"
                        }
                    },
                    "400": {
                        "description": "Datos enviados incorrectos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
        "/api/docentes": {
            "get": {
                "tags": [
                    {
                        "name": "Docente",
                        "description": " API para los docentes"
                    }
                ],
                "summary": "Obtiene todos los docentes",
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa",
                        "schema": {
                            "$ref": "#/definitions/Docente"
                        }
                    },
                    "400": {
                        "description": "Datos enviados incorrectos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "post": {
                "tags": [
                    {
                        "name": "Docente",
                        "description": " API para los docentes"
                    }
                ],
                "summary": "Crea un docente",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "El docente que se creará, no debe incluir id",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Docente"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa",
                        "schema": {
                            "$ref": "#/definitions/Docente"
                        }
                    },
                    "400": {
                        "description": "Datos enviados incorrectos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
        "/api/docentes/:id": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID del docente",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    {
                        "name": "Docente",
                        "description": " API para los docentes"
                    }
                ],
                "summary": "Obtiene el docente con el ID dado",
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa",
                        "schema": {
                            "$ref": "#/definitions/Docente"
                        }
                    },
                    "400": {
                        "description": "Datos enviados incorrectos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    {
                        "name": "Docente",
                        "description": " API para los docentes"
                    }
                ],
                "summary": "Actualiza al docente con el ID dado",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "description": "El docente se actulizará, no debe incluir CI",
                        "schema": {
                            "$ref": "#/definitions/Docente"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa",
                        "schema": {
                            "$ref": "#/definitions/Docente"
                        }
                    },
                    "400": {
                        "description": "Datos enviados incorrectos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    {
                        "name": "Docente",
                        "description": " API para los docentes"
                    }
                ],
                "summary": "Elimina al docente con el ID dado",
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa",
                        "schema": {
                            "$ref": "#/definitions/Docente"
                        }
                    },
                    "400": {
                        "description": "Datos enviados incorrectos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
        "/api/materias": {
            "get": {
                "tags": [
                    {
                        "name": "Materia",
                        "description": " API para las materias"
                    }
                ],
                "summary": "Obtiene todos las materias",
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa",
                        "schema": {
                            "$ref": "#/definitions/Materia"
                        }
                    },
                    "400": {
                        "description": "Datos enviados incorrectos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
        "/api/materias/:id": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID de la materia que se quiere obtener",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    {
                        "name": "Materia",
                        "description": " API para las materias"
                    }
                ],
                "summary": "Obtiene la materia con el ID dado",
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa",
                        "schema": {
                            "$ref": "#/definitions/Materia"
                        }
                    },
                    "400": {
                        "description": "Datos enviados incorrectos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
        "/api/estudiantes/:id/materias": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID del estudiante del que se quiere obtener materias y notas",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    {
                        "name": "Materias y notas",
                        "description": " API para las materias y notas"
                    }
                ],
                "summary": "Obtiene las materias y notas con el ID dado",
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa",
                        "schema": {
                            "$ref": "#/definitions/Estudiante_has_materia"
                        }
                    },
                    "400": {
                        "description": "Datos enviados incorrectos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
        "/api/jefeCarrera/inscripcion": {
            "post": {
                "tags": [
                    {
                        "name": "Asignación",
                        "description": " API para las asignaciones"
                    }
                ],
                "summary": "Asigna un estudiante a una materia",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "El estudiante sera asignado, no debe incluir ninguna nota",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Estudiante_has_materia"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa",
                        "schema": {
                            "$ref": "#/definitions/Estudiante_has_materia"
                        }
                    },
                    "400": {
                        "description": "Datos enviados incorrectos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
        "/api/jefeCarrera/asignacion": {
            "post": {
                "tags": [
                    {
                        "name": "Asignación",
                        "description": " API para las asignaciones"
                    }
                ],
                "summary": "Asigna un docente a una materia",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "El docente sera asignado",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Docente_has_materia"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa",
                        "schema": {
                            "$ref": "#/definitions/Docente_has_materia"
                        }
                    },
                    "400": {
                        "description": "Datos enviados incorrectos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
...
        "/api/docentes/:id/materias": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID del docente del que se quiere obtener materias",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    {
                        "name": "Docentes y materias",
                        "description": " API para los docentes y materias"
                    }
                ],
                "summary": "Obtiene las materias que un docente dicta el ID dado",
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa",
                        "schema": {
                            "$ref": "#/definitions/Docente_has_materia"
                        }
                    },
                    "400": {
                        "description": "Datos enviados incorrectos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
        "/api/kardex/materias": {
            "get": {
                "tags": [
                    {
                        "name": "Docentes y materias",
                        "description": " API para los docentes y materias"
                    }
                ],
                "summary": "Obtiene las materias que los docentes dictan",
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa",
                        "schema": {
                            "$ref": "#/definitions/Docente_has_materia"
                        }
                    },
                    "400": {
                        "description": "Datos enviados incorrectos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Estudiante": {
            "properties": {
                "idEstudiante": {
                    "type": "integer"
                },
                "nombre": {
                    "type": "string"
                },
                "apellido_1": {
                    "type": "string"
                },
                "apellido_2": {
                    "type": "string"
                },
                "ci": {
                    "type": "integer"
                },
                "fecha_nacimiento": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "Docente": {
            "properties": {
                "idDocente": {
                    "type": "integer"
                },
                "nombre": {
                    "type": "string"
                },
                "apellido_1": {
                    "type": "string"
                },
                "apellido_2": {
                    "type": "string"
                },
                "ci": {
                    "type": "integer"
                },
                "fecha_nacimiento": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
    }
}